+++ packages/oft-hyperliquid-evm/src/signer/wallet.ts
@@ -1,17 +1,20 @@
-import { Wallet } from 'ethers'
-import axios, { AxiosInstance } from 'axios'
-import { HYPERLIQUID_URLS } from '@/types'
-
 import { getTimestampMs, signL1Action } from '@/signer'
+import { HYPERLIQUID_URLS, ValueType } from '@/types'
+import axios, { AxiosInstance } from 'axios'
+import { Wallet } from 'ethers'
+
+import { Logger, createLogger } from '@layerzerolabs/io-devtools'
 
 export class HyperliquidClient {
     private readonly client: AxiosInstance
     private readonly baseUrl: string
     private readonly isTestnet: boolean
+    private readonly logger: Logger
 
-    constructor(isTestnet: boolean) {
+    constructor(isTestnet: boolean, logLevel: string) {
         this.baseUrl = isTestnet ? HYPERLIQUID_URLS.TESTNET : HYPERLIQUID_URLS.MAINNET
         this.isTestnet = isTestnet
+        this.logger = createLogger(logLevel)
         this.client = axios.create({
             baseURL: this.baseUrl,
             headers: {
@@ -20,7 +23,7 @@ export class HyperliquidClient {
         })
     }
 
-    async submitHyperliquidAction(endpoint: string, wallet: Wallet, action: any) {
+    async submitHyperliquidAction(endpoint: string, wallet: Wallet, action: ValueType) {
         const nonce = getTimestampMs()
 
         const signature = await signL1Action({
@@ -39,17 +42,19 @@ export class HyperliquidClient {
             vaultAddress: null,
         }
 
-        try {
-            console.log('Sending payload:', JSON.stringify(payload, null, 2))
-            console.log(`Sending to: ${this.baseUrl}${endpoint}`)
+        this.logger.debug(`Sending payload to full url: ${this.baseUrl}${endpoint}`)
+        this.logger.debug(`payload: ${JSON.stringify(payload, null, 2)}`)
 
-            // Direct POST to the full endpoint
+        try {
             const response = await this.client.post(endpoint, payload)
             return response.data
         } catch (error) {
             console.error(error)
+            this.logger.error(error)
+            this.logger.error(`Sending payload to full url: ${this.baseUrl}${endpoint}`)
+            this.logger.error(`payload: ${JSON.stringify(payload, null, 2)}`)
             if (axios.isAxiosError(error) && error.response) {
-                console.error('API Error:', {
+                this.logger.error('API Error:', {
                     status: error.response.status,
                     statusText: error.response.statusText,
                     data: error.response.data,