+++ packages/omnicounter-devtools/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @layerzerolabs/omnicounter-devtools
 
+## 3.0.6
+
+### Patch Changes
+
+- 8b6c422: Bump monorepo dependencies to latest patch version
+- Updated dependencies [8b6c422]
+  - @layerzerolabs/protocol-devtools@1.1.6
+  - @layerzerolabs/devtools@0.4.8
+
 ## 3.0.5
 
 ### Patch Changes